copy all resources after write
